// src/app/model/game-state.enum.ts
export enum GameState {
  WAITING = 'waiting',
  // SELECTED = 'selected',
  // OPENED = 'opened',
  CLOSED = 'closed',
  WIN = 'win',

  MATCH = 'match',
  MISSMATCH = 'missmatch',

  SELECTED_FIRST = 'selected_first',
  SELECTED_SECOND = 'selected_second',
  OPENED_FIRST = 'opened_first',
  OPENED_SECOND = 'opened_second',
}